server/requirements.txt
autopep8==1.5.7
cffi==1.14.5
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
Flask==2.0.1
greenlet==1.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.1.0
PyMySQL==1.0.2
SQLAlchemy==1.4.18
toml==0.10.2
Werkzeug==2.0.1